Refactored group service to filter closed groups & refined user role checks
The changes on group.service.ts added additional filter conditions to skip groups with a 'CLOSED' status when users list 'PUBLIC' or 'PRIVATE' groups.

The 'ifAccountAdminOfGroup' method was renamed to 'isMemberGroupAdmin', fetching whether the member has admin privileges regardless of the group type. This modification simplifies the retrieval process and ensures the user role is correctly identified across both 'PUBLIC' and 'PRIVATE' groups.

For the 'closeGroup' method, the admin check was removed as only admins can close groups. This is done externally via the resolver.

In group.resolver.ts, 'checkIfMemberIsGroupAdmin' was added to ensure that only an admin user can update or delete a group.

The 'event-application-pair.resolver.ts' file was also updated to replace 'ifAccountAdminOfGroup' method with the new 'isMemberGroupAdmin'.

+++ src/app/group/group.resolver.ts
@@ -62,6 +62,20 @@ export class GroupResolver {
     }
   }
 
+  private async checkIfMemberIsGroupAdmin(
+    accountId: number,
+    groupId: number,
+  ): Promise<void> {
+    const isAdmin = await this.groupService.isMemberGroupAdmin(
+      accountId,
+      groupId,
+    );
+
+    if (!isAdmin) {
+      throw new Error(this.i18n.t('errors.unauthorized'));
+    }
+  }
+
   @Query(() => [Group], { name: 'publicGroups' })
   @UseGuards(AuthGuard)
   async publicGroups(
@@ -112,6 +126,7 @@ export class GroupResolver {
     @Args('input') input: CreateOrUpdateGroupInput,
   ): Promise<Group> {
     await this.checkGroupMembership(context.account!.id, id);
+    await this.checkIfMemberIsGroupAdmin(context.account!.id, id);
     return this.groupService.updateGroup(context.account!.id, id, input);
   }
 
@@ -119,9 +134,10 @@ export class GroupResolver {
   @UseGuards(AuthGuard)
   async deleteGroup(
     @RequestContextDecorator() context: RequestContext,
-    @Args('id') id: number,
+    @Args('id', { type: () => Int }) id: number,
   ): Promise<Group> {
     await this.checkGroupMembership(context.account!.id, id);
+    await this.checkIfMemberIsGroupAdmin(context.account!.id, id);
     return this.groupService.deleteGroup(context.account!.id, id);
   }
